move gamma flattening out of geocoding and apps.
