+++ sarsen/scene.py
@@ -1,6 +1,7 @@
 import typing as T
 
 import numpy as np
+import numpy.typing as npt
 import xarray as xr
 from rasterio import warp
 
@@ -55,3 +56,48 @@ def transform_dem_3d(
 def convert_to_dem_ecef(dem_raster: xr.DataArray) -> xr.DataArray:
     dem_3d = convert_to_dem_3d(dem_raster)
     return transform_dem_3d(dem_3d, target_crs=ECEF_CRS)
+
+
+def compute_dem_oriented_area(
+        dem_ecef: xr.DataArray,
+) -> xr.DataArray:
+    x_corners: npt.ArrayLike = np.concatenate(
+        [
+            [dem_ecef.x[0] + (dem_ecef.x[0] - dem_ecef.x[1]) / 2],
+            ((dem_ecef.x.shift(x=-1) + dem_ecef.x) / 2)[:-1].data,
+            [dem_ecef.x[-1] + (dem_ecef.x[-1] - dem_ecef.x[-2]) / 2],
+        ]
+    )
+    y_corners: npt.ArrayLike = np.concatenate(
+        [
+            [dem_ecef.y[0] + (dem_ecef.y[0] - dem_ecef.y[1]) / 2],
+            ((dem_ecef.y.shift(y=-1) + dem_ecef.y) / 2)[:-1].data,
+            [dem_ecef.y[-1] + (dem_ecef.y[-1] - dem_ecef.y[-2]) / 2],
+        ]
+    )
+    dem_ecef_corners = dem_ecef.interp(
+        {"x": x_corners, "y": y_corners},
+        method="linear",
+        kwargs={"fill_value": "extrapolate"},
+    )
+
+    dx = dem_ecef_corners.diff("x", 1)
+    dy = dem_ecef_corners.diff("y", 1)
+
+    dx1 = dx.isel(y=slice(1, None)).assign_coords(dem_ecef.coords)  # type: ignore
+    dy1 = dy.isel(x=slice(1, None)).assign_coords(dem_ecef.coords)  # type: ignore
+    dx2 = dx.isel(y=slice(None, -1)).assign_coords(dem_ecef.coords)  # type: ignore
+    dy2 = dy.isel(x=slice(None, -1)).assign_coords(dem_ecef.coords)  # type: ignore
+
+    cross_1 = xr.cross(dx1, dy1, dim="axis") / 2
+    sign_1 = np.sign(
+        xr.dot(cross_1, dem_ecef, dims="axis")  # type: ignore
+    )  # ensure direction out of DEM
+
+    cross_2 = xr.cross(dx2, dy2, dim="axis") / 2
+    sign_2 = np.sign(
+        xr.dot(cross_2, dem_ecef, dims="axis")  # type: ignore
+    )  # ensure direction out of DEM
+    dem_oriented_area: xr.DataArray = cross_1 * sign_1 + cross_2 * sign_2
+
+    return dem_oriented_area
